multi-node clusters can now be formed
membership views are incrementally created, modifying the existing view
with the new changes (ups & downs).
problem: this does not satisfy the requirements of the distributed lock
service to have an ordered list of members that is the same in each node

CONTINUING PROBLEM: I'm using the configuration identifier (a Long) as
the view ID.  This number is not a a Lamport clock value like it is in
JGroups, so we can't distinguish the order of views.  Some code checks
for this, such as code determining whether an old process that's been
kicked out of the cluster is hanging around trying to communicate with
members that are still in the cluster.  The "birth view ID" was being
used for this check.  Also, under JGroups we could reject a new view as
being older than the one we already had based on the view ID being a
Lamport clock value (monotonically increasing in value).